sessions: test: restore focused test modules
Keep the sessions tests organized by concern because the Windows EMFILE failure is repository-wide and unrelated to this change.


Copilot-Session: 0cd76124-732c-4db6-9c17-0e251d10dd10
